move modules to main module folder
