+++ client/src/components/Workflow/Editor/modules/onAllInputs.ts
