+++ client/src/components/Workflow/Editor/modules/removeOpenConnections.ts
